Fixnuty error pri kotrolovani presvihnutych chemikalii, rovnako tak fixnuty error, pri validacii desatinnych cisel

+++ skladisko/app/assets/javascripts/application.js
@@ -318,7 +318,7 @@ $(document).ready( function () {
 		array.eq(i).css({
 			'border': '2px solid black'	
 		});
-		if (!$.isNumeric(array.eq(i).val()) || parseInt(array.eq(i).val()) <= 0) {
+		if (!$.isNumeric(array.eq(i).val()) || parseFloat(array.eq(i).val()) <= 0) {
 			array.eq(i).css({
 				'border': '2px solid red'	
 			});
